+++ src/models/models.py
@@ -6,24 +6,26 @@ from src.database import db
 
 
 class User(db.Model):
-    __tablename__ = 'users'
-    __table_args__ = {'extend_existing': True}
+    __tablename__ = "users"
+    __table_args__ = {"extend_existing": True}
 
     user_id = db.Column(db.Integer, primary_key=True, unique=True)
     name = db.Column(db.Unicode(255), nullable=False)
     created_at = db.Column(db.DateTime, nullable=False, default=datetime.now)
-    updated_at = db.Column(db.DateTime, nullable=False, default=datetime.now, onupdate=datetime.now)
+    updated_at = db.Column(
+        db.DateTime, nullable=False, default=datetime.now, onupdate=datetime.now
+    )
     session_id = db.Column(db.Unicode(255))
 
     # likeunlike = db.relationship('LikeUnlike', backref='users')
 
     def __repr__(self):
-        return '<User id={id} name={name}>'.format(
-                id=self.id, name=self.name)
+        return "<User id={id} name={name}>".format(id=self.id, name=self.name)
+
 
 class AnimeData(db.Model):
-    __tablename__ = 'anime_data'
-    __table_args__ = {'extend_existing': True}
+    __tablename__ = "anime_data"
+    __table_args__ = {"extend_existing": True}
 
     anime_id = db.Column(db.Integer, primary_key=True, unique=True)
     title = db.Column(db.Unicode(255), nullable=False)
@@ -35,25 +37,34 @@ class AnimeData(db.Model):
 
     # likeunlike = db.relationship('LikeUnlike', backref='anime_data')
 
+
 class LikeUnlike(db.Model):
-    __tablename__ = 'likeunlike'
-    __table_args__ = {'extend_existing': True}
+    __tablename__ = "likeunlike"
+    __table_args__ = {"extend_existing": True}
 
     id = db.Column(db.Integer, primary_key=True, unique=True)
-    #user_id = db.Column(db.Integer, db.ForeignKey('users.user_id'), nullable=False)
+    # user_id = db.Column(db.Integer, db.ForeignKey('users.user_id'), nullable=False)
     user_id = db.Column(db.Integer, nullable=False)
-    #anime_id = db.Column(db.Integer, db.ForeignKey('anime_data.anime_id'), nullable=False)
+    # anime_id = db.Column(db.Integer,
+    #                      db.ForeignKey('anime_data.anime_id'),
+    #                      nullable=False
+    #                     )
     anime_id = db.Column(db.Integer, nullable=False)
     status = db.Column(db.Integer, nullable=False)
     created_at = db.Column(db.DateTime, nullable=False, default=datetime.now)
-    updated_at = db.Column(db.DateTime, nullable=False, default=datetime.now, onupdate=datetime.now)
+    updated_at = db.Column(
+        db.DateTime, nullable=False, default=datetime.now, onupdate=datetime.now
+    )
+
 
 class Recommended(db.Model):
-    __tablename__ = 'recommended'
-    __table_args__ = {'extend_existing': True}
+    __tablename__ = "recommended"
+    __table_args__ = {"extend_existing": True}
 
     id = db.Column(db.Integer, primary_key=True, unique=True)
     user_id = db.Column(db.Integer, nullable=False)
     anime_id = db.Column(db.Integer, nullable=False)
     created_at = db.Column(db.DateTime, nullable=False, default=datetime.now)
-    updated_at = db.Column(db.DateTime, nullable=False, default=datetime.now, onupdate=datetime.now)
+    updated_at = db.Column(
+        db.DateTime, nullable=False, default=datetime.now, onupdate=datetime.now
+    )
