+++ src/requirements.txt
@@ -1,4 +1,5 @@
 flask
+flask-restful
 sqlalchemy<1.4
 flask-sqlalchemy
 sqlalchemy_utils
